extract d3 decorators to separate file
